+++ vowpalwabbit/lda_core.cc
@@ -854,7 +854,7 @@ std::istream &operator>>(std::istream &in, lda_math_mode &mmode)
 }
 
 LEARNER::base_learner *lda_setup(vw &all)
-{ if (missing_option<uint64_t, true>(all, "lda", "Run lda with <int> topics"))
+{ if (missing_option<uint32_t, true>(all, "lda", "Run lda with <int> topics"))
     return nullptr;
   new_options(all, "Lda options")
     ("lda_alpha", po::value<float>()->default_value(0.1f),"Prior on sparsity of per-document topic weights")
